# openapi_check.json
{"openapi":"3.1.0","info":{"title":"투자만능계산기 API","version":"6.0.0"},"paths":{"/api/v1/stocks/exchange-rate":{"get":{"tags":["stocks"],"summary":"Get Exchange Rate","operationId":"get_exchange_rate_api_v1_stocks_exchange_rate_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExchangeRateResponse"}}}}}}},"/api/v1/stocks/search":{"get":{"tags":["stocks"],"summary":"Search Stocks","operationId":"search_stocks_api_v1_stocks_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stocks/{ticker}":{"get":{"tags":["stocks"],"summary":"Get Stock","operationId":"get_stock_api_v1_stocks__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stocks/{ticker}/history":{"get":{"tags":["stocks"],"summary":"Get Stock History","operationId":"get_stock_history_api_v1_stocks__ticker__history_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"period","in":"query","required":false,"schema":{"type":"string","default":"1y","title":"Period"}},{"name":"interval","in":"query","required":false,"schema":{"type":"string","default":"1d","title":"Interval"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backtests/infinite-buy":{"post":{"tags":["backtests"],"summary":"Run Infinite Buy","operationId":"run_infinite_buy_api_v1_backtests_infinite_buy_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InfiniteBuyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backtests/dca":{"post":{"tags":["backtests"],"summary":"Run Dca","operationId":"run_dca_api_v1_backtests_dca_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DCARequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backtests/buy-and-hold":{"post":{"tags":["backtests"],"summary":"Run Buy And Hold","operationId":"run_buy_and_hold_api_v1_backtests_buy_and_hold_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyAndHoldRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backtests/value-rebalance":{"post":{"tags":["backtests"],"summary":"Run Value Rebalance","operationId":"run_value_rebalance_api_v1_backtests_value_rebalance_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValueRebalanceRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backtests/compare":{"post":{"tags":["backtests"],"summary":"Compare Strategies","operationId":"compare_strategies_api_v1_backtests_compare_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompareRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backtests/presets":{"get":{"tags":["backtests"],"summary":"Get Presets","operationId":"get_presets_api_v1_backtests_presets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BacktestConfigSchema":{"properties":{"buy_commission":{"type":"number","title":"Buy Commission","default":0.0},"sell_commission":{"type":"number","title":"Sell Commission","default":0.0},"slippage":{"type":"number","title":"Slippage","default":0.0},"fx_spread":{"type":"number","title":"Fx Spread","default":0.0},"capital_gains_tax":{"type":"number","title":"Capital Gains Tax","default":0.0},"dividend_tax":{"type":"number","title":"Dividend Tax","default":0.0},"allow_fractional":{"type":"boolean","title":"Allow Fractional","default":true},"preset":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Preset"}},"type":"object","title":"BacktestConfigSchema"},"BuyAndHoldRequest":{"properties":{"ticker":{"type":"string","title":"Ticker"},"capital":{"type":"number","title":"Capital"},"period":{"type":"string","title":"Period","default":"1y"},"config":{"$ref":"#/components/schemas/BacktestConfigSchema","default":{"buy_commission":0.0,"sell_commission":0.0,"slippage":0.0,"fx_spread":0.0,"capital_gains_tax":0.0,"dividend_tax":0.0,"allow_fractional":true}}},"type":"object","required":["ticker","capital"],"title":"BuyAndHoldRequest"},"CompareRequest":{"properties":{"ticker":{"type":"string","title":"Ticker"},"capital":{"type":"number","title":"Capital"},"period":{"type":"string","title":"Period","default":"1y"},"strategies":{"items":{"type":"string"},"type":"array","title":"Strategies","default":["infinite_buy","dca","buy_and_hold","value_rebalance"]},"config":{"$ref":"#/components/schemas/BacktestConfigSchema","default":{"buy_commission":0.0,"sell_commission":0.0,"slippage":0.0,"fx_spread":0.0,"capital_gains_tax":0.0,"dividend_tax":0.0,"allow_fractional":true}}},"type":"object","required":["ticker","capital"],"title":"CompareRequest"},"DCARequest":{"properties":{"ticker":{"type":"string","title":"Ticker"},"capital":{"type":"number","title":"Capital"},"periodic_amount":{"type":"number","title":"Periodic Amount"},"interval":{"type":"integer","title":"Interval","default":21},"initial_lump":{"type":"number","title":"Initial Lump","default":0},"period":{"type":"string","title":"Period","default":"1y"},"config":{"$ref":"#/components/schemas/BacktestConfigSchema","default":{"buy_commission":0.0,"sell_commission":0.0,"slippage":0.0,"fx_spread":0.0,"capital_gains_tax":0.0,"dividend_tax":0.0,"allow_fractional":true}}},"type":"object","required":["ticker","capital","periodic_amount"],"title":"DCARequest"},"ExchangeRateResponse":{"properties":{"rate":{"type":"number","title":"Rate"},"currency_pair":{"type":"string","title":"Currency Pair","default":"USD/KRW"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["rate","timestamp"],"title":"ExchangeRateResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InfiniteBuyRequest":{"properties":{"ticker":{"type":"string","title":"Ticker"},"capital":{"type":"number","title":"Capital"},"divisions":{"type":"integer","title":"Divisions","default":40},"target_profit":{"type":"number","title":"Target Profit","default":10.0},"period":{"type":"string","title":"Period","default":"1y"},"config":{"$ref":"#/components/schemas/BacktestConfigSchema","default":{"buy_commission":0.0,"sell_commission":0.0,"slippage":0.0,"fx_spread":0.0,"capital_gains_tax":0.0,"dividend_tax":0.0,"allow_fractional":true}}},"type":"object","required":["ticker","capital"],"title":"InfiniteBuyRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"ValueRebalanceRequest":{"properties":{"ticker":{"type":"string","title":"Ticker"},"capital":{"type":"number","title":"Capital"},"upper_band":{"type":"number","title":"Upper Band","default":1.15},"lower_band":{"type":"number","title":"Lower Band","default":0.85},"annual_growth":{"type":"number","title":"Annual Growth","default":0.15},"pool_limit":{"type":"number","title":"Pool Limit","default":0.25},"period":{"type":"string","title":"Period","default":"1y"},"config":{"$ref":"#/components/schemas/BacktestConfigSchema","default":{"buy_commission":0.0,"sell_commission":0.0,"slippage":0.0,"fx_spread":0.0,"capital_gains_tax":0.0,"dividend_tax":0.0,"allow_fractional":true}}},"type":"object","required":["ticker","capital"],"title":"ValueRebalanceRequest"}}}}